+++ lib/ts/recipe/emailpassword/componentOverrideContext.tsx
@@ -1,10 +1,10 @@
 import { createGenericComponentsOverrideContext } from "../../components/componentOverride/genericComponentOverrideContext";
 import { ComponentOverrideMap } from "./types";
 
-const [useOverrideContext, Provider, Consumer] = createGenericComponentsOverrideContext<ComponentOverrideMap>();
+const [useContext, Provider, Consumer] = createGenericComponentsOverrideContext<ComponentOverrideMap>();
 
 export {
-    useOverrideContext,
+    useContext as useRecipeComponentOverrideContext,
     Provider as RecipeComponentsOverrideContextProvider,
     Consumer as RecipeComponentsOverrideContextConsumer,
 };