+++ lib/ts/recipe/thirdparty/componentOverrideContext.tsx
@@ -1,6 +1,6 @@
 import { createGenericComponentsOverrideContext } from "../../components/componentOverride/genericComponentOverrideContext";
 import { ComponentOverrideMap } from "./types";
 
-const [useOverrideContext, Provider] = createGenericComponentsOverrideContext<ComponentOverrideMap>();
+const [useContext, Provider] = createGenericComponentsOverrideContext<ComponentOverrideMap>();
 
-export { useOverrideContext, Provider as RecipeComponentsOverrideContextProvider };
+export { useContext as useRecipeComponentOverrideContext, Provider as RecipeComponentsOverrideContextProvider };